TSx

Documentation for TSx.

TSx.JoinBoth

TSx.TS

TSx.JoinBoth — Type

Joins/Column-binding

Joins/Column-binding

TS objects can be combined together column-wise using Index as the column key. There are four kinds of column-binding operations possible as of now. Each join operation works by performing a Set operation on the Index column and then merging the datasets based on the output from the Set operation. Each operation changes column names in the final object automatically if the operation encounters duplicate column names amongst the TS objects.

The following join types are supported:

join(ts1::TS, ts2::TS, ::JoinBoth): a.k.a. inner join, takes the

intersection of the indexes of ts1 and ts2, and then merges the columns of both the objects. The resulting object will only contain rows which are present in both the objects' indexes. The function will renamine the columns in the final object if they had same names in the TS objects.

join(ts1::TS, ts2::TS, ::JoinAll): a.k.a. outer join, takes the

union of the indexes of ts1 and ts2 before merging the other columns of input objects. The output will contain rows which are present in all the input objects while inserting missing values where a row was not present in any of the objects. This is the default behaviour if no JoinType object is provided.

join(ts1::TS, ts2::TS, ::JoinLeft): a.k.a. left join, takes the

index values which are present in the left object ts1 and finds matching index values in the right object ts2. The resulting object includes all the rows from the left object, the column values from the left object, and the values associated with matching index rows on the right. The operation inserts missing values where in the unmatched rows of the right object.

join(ts1::TS, ts2::TS, ::JoinRight): a.k.a. left join, is similar

to left join but works in the opposite direction. The final object contains all the rows from the right object while inserting missing values in rows missing from the left object.

The default behaviour is to assume JoinAll() if no JoinType object is provided to the join method.

Examples

julia> df1 = DataFrame(Index=1:10, x1 = randn(10))
julia> ts1 = TS(df1)
julia> df2 = DataFrame(Index=1:10, x2 = randn(10))
julia> ts2 = TS(df2)

julia> join(ts1, ts2, JoinAll()) # with `missing` inserted
# same as JoinAll()
julia> join(ts1, ts2)            
julia> join(ts1, ts2, JoinBoth())
julia> join(ts1, ts2, JoinLeft())
julia> join(ts1, ts2, JoinRight())

# Using TimeType objects
julia> dates = collect(Date(2017,1,1):Day(1):Date(2017,1,10))
julia> ts1 = TS(DataFrame(Index = dates, x1 = randn(10)))
julia> dates = collect(Date(2017,1,1):Day(1):Date(2017,1,30))
julia> ts2 = TS(DataFrame(Index = dates, x1 = randn(30)))

julia> join(ts1, ts2)

TSx.TS — Type

TS

A type to hold ordered data with an index.

A TS object is essentially a DataFrame with a specific column marked as an index and has the name Index. The DataFrame is sorted using index values during construction.

Constructors

TS(coredata::DataFrame, index::Union{String, Symbol, Int}=1)
TS(coredata::DataFrame, index::AbstractVector{T}) where {T<:Int}
TS(coredata::DataFrame, index::UnitRange{Int})
TS(coredata::AbstractVector{T}, index::AbstractVector{V}) where {T, V}
TS(coredata::AbstractVector{T}) where {T}
TS(coredata::AbstractArray{T,2}, meta::Dict=Dict{String, Any}()) where {T}

Examples

julia> df = DataFrame(x1 = randn(10))
julia> TS(df)

julia> df = DataFrame(Index = [1, 2, 3], x1 = randn(3))
julia> TS(df, 1)

julia> dates = collect(Date(2017,1,1):Day(1):Date(2017,1,10))
julia> df = DataFrame(dates = dates, x1 = randn(10))
julia> TS(df, :dates)
julia> TS(DataFrame(x1=randn(10), dates))

julia> TS(randn(10))
julia> TS(randn(10), dates)