TSx

Documentation for TSx.

TSx.TS

Base.getindex

Base.join

Base.names

Base.size

Base.vcat

RecipesBase.apply_recipe

TSx.apply

TSx.index

TSx.lag

TSx.lead

TSx.log

TSx.ncol

TSx.nrow

TSx.rollapply

TSx.TS — Type

struct TS
  coredata :: DataFrame
end

::TS - A type to hold ordered data with an index.

A TS object is essentially a DataFrame with a specific column marked as an index. The input DataFrame is sorted during construction and is stored under the property coredata. The index is stored in the Index column of coredata.

Permitted data inputs to the constructors are DataFrame, Vector, and 2-dimensional Array. If an index is already not present in the constructor then a sequential integer index is created automatically.

TS(coredata::DataFrame): Here, the constructor looks for a column named Index in coredata as the index column, if this is not found then the first column of coredata is made the index by default. If coredata only has a single column then a new sequential index is generated.

Since TS.coredata is a DataFrame it can be operated upon independently using methods provided by the DataFrames package (ex. transform, combine, etc.).

Constructors

TS(coredata::DataFrame, index::Union{String, Symbol, Int})
TS(coredata::DataFrame, index::AbstractVector{T}) where {T<:Union{Int, TimeType}}
TS(coredata::DataFrame)
TS(coredata::DataFrame, index::UnitRange{Int})
TS(coredata::AbstractVector{T}, index::AbstractVector{V}) where {T, V}
TS(coredata::AbstractVector{T}) where {T}
TS(coredata::AbstractArray{T,2}) where {T}
TS(coredata::AbstractArray{T,2}, index::AbstractVector{V}) where {T, V}

Examples

julia> using Random;
julia> random(x) = rand(MersenneTwister(123), x);

julia> df = DataFrame(x1 = random(10))

julia> ts = TS(df)   # generates index
(10 x 1) TS with Int64 Index

 Index  x1        
 Int64  Float64   
──────────────────
     1  0.768448
     2  0.940515
     3  0.673959
     4  0.395453
     5  0.313244
     6  0.662555
     7  0.586022
     8  0.0521332
     9  0.26864
    10  0.108871

# ts.coredata is a DataFrame
julia> combine(ts.coredata, :x1 => Statistics.mean, DataFrames.nrow)
1×2 DataFrame
 Row │ x1_mean  nrow
     │ Float64  Int64
─────┼────────────────
   1 │ 0.49898    418

julia> df = DataFrame(ind = [1, 2, 3], x1 = random(3))
3×2 DataFrame
 Row │ ind    x1       
     │ Int64  Float64  
─────┼─────────────────
   1 │     1  0.768448
   2 │     2  0.940515
   3 │     3  0.673959

julia> ts = TS(df, 1)        # the first column is index
(3 x 1) TS with Int64 Index

 Index  x1       
 Int64  Float64  
─────────────────
     1  0.768448
     2  0.940515
     3  0.673959

julia> df = DataFrame(x1 = random(3), x2 = random(3), Index = [1, 2, 3]);
3×3 DataFrame
 Row │ x1        x2        Index 
     │ Float64   Float64   Int64 
─────┼───────────────────────────
   1 │ 0.768448  0.768448      1
   2 │ 0.940515  0.940515      2
   3 │ 0.673959  0.673959      3

julia> ts = TS(df)   # uses existing `Index` column
(3 x 2) TS with Int64 Index

 Index  x1        x2       
 Int64  Float64   Float64  
───────────────────────────
     1  0.768448  0.768448
     2  0.940515  0.940515
     3  0.673959  0.673959

julia> dates = collect(Date(2017,1,1):Day(1):Date(2017,1,10));

julia> df = DataFrame(dates = dates, x1 = random(10))

julia> ts = TS(df, :dates)
(10 x 1) TS with Date Index

 Index       x1        
 Date        Float64   
───────────────────────
 2017-01-01  0.768448
 2017-01-02  0.940515
 2017-01-03  0.673959
 2017-01-04  0.395453
 2017-01-05  0.313244
 2017-01-06  0.662555
 2017-01-07  0.586022
 2017-01-08  0.0521332
 2017-01-09  0.26864
 2017-01-10  0.108871

julia> ts = TS(DataFrame(x1=random(10)), dates)

julia> ts = TS(random(10))
(10 x 1) TS with Int64 Index

 Index  x1        
 Int64  Float64   
──────────────────
     1  0.768448
     2  0.940515
     3  0.673959
     4  0.395453
     5  0.313244
     6  0.662555
     7  0.586022
     8  0.0521332
     9  0.26864
    10  0.108871

julia> ts = TS(random(10), dates);

julia> ts = TS([random(10) random(10)], dates) # matrix object
(10 x 2) TS with Date Index

 Index       x1         x2        
 Date        Float64    Float64   
──────────────────────────────────
 2017-01-01  0.768448   0.768448
 2017-01-02  0.940515   0.940515
 2017-01-03  0.673959   0.673959
 2017-01-04  0.395453   0.395453
 2017-01-05  0.313244   0.313244
 2017-01-06  0.662555   0.662555
 2017-01-07  0.586022   0.586022
 2017-01-08  0.0521332  0.0521332
 2017-01-09  0.26864    0.26864
 2017-01-10  0.108871   0.108871

Base.getindex — Method

Subsetting/Indexing

TS can be subset using row and column indices. The row selector could be an integer, a range, an array or it could also be a Date object or an ISO-formatted date string ("2007-04-10"). There are methods to subset on year, year-month, and year-quarter.

The latter two subset coredata by matching on the index column.

Column selector could be an integer or any other selector which DataFrame indexing supports. To fetch the index column one can use the index() method on the TS object.

Examples

julia> using Random;

julia> random(x) = rand(MersenneTwister(123), x);

julia> ts = TS([random(10) random(10) random(10)])

# first row
julia> ts[1]            |> print
(1 x 3) TS with Int64 Index

 Index  x1        x2        x3
 Int64  Float64   Float64   Float64
─────────────────────────────────────
     1  0.768448  0.768448  0.768448

# first five rows
julia> ts[1:5]          |> print
(5 x 3) TS with Int64 Index

 Index  x1        x2        x3
 Int64  Float64   Float64   Float64
─────────────────────────────────────
     1  0.768448  0.768448  0.768448
     2  0.940515  0.940515  0.940515
     3  0.673959  0.673959  0.673959
     4  0.395453  0.395453  0.395453
     5  0.313244  0.313244  0.313244

# first five rows, second column
julia> ts[1:5, 2]       |> print
(5 x 1) TS with Int64 Index

 Index  x2
 Int64  Float64
─────────────────
     1  0.768448
     2  0.940515
     3  0.673959
     4  0.395453
     5  0.313244

julia> ts[1:5, 2:3]     |> print

# individual rows
julia> ts[[1, 9]]     |> print

julia> dates = collect(Date(2007):Day(1):Date(2008, 2, 22));

julia> ts = TS(random(length(dates)), dates);

julia> ts[Date(2007, 01, 01)]           |> print
(1 x 1) TS with Dates.Date Index

 Index       x1
 Date        Float64
──────────────────────
 2007-01-01  0.768448

julia> ts[Date(2007)]           |> print
(1 x 1) TS with Dates.Date Index

 Index       x1
 Date        Float64
──────────────────────
 2007-01-01  0.768448

julia> ts[Year(2007)]           |> print
(365 x 1) TS with Dates.Date Index

 Index       x1
 Date        Float64
───────────────────────
 2007-01-01  0.768448
 2007-01-02  0.940515
 2007-01-03  0.673959
 2007-01-04  0.395453
 2007-01-05  0.313244
 2007-01-06  0.662555
 2007-01-07  0.586022
 2007-01-08  0.0521332
     ⋮           ⋮
 2007-12-24  0.468421
 2007-12-25  0.0246652
 2007-12-26  0.171042
 2007-12-27  0.227369
 2007-12-28  0.695758
 2007-12-29  0.417124
 2007-12-30  0.603757
 2007-12-31  0.346659
       349 rows omitted

julia> ts[Year(2007), Month(11)];

julia> ts[Year(2007), Quarter(2)];

julia> ts["2007-01-01"]

julia> ts[1, :x1]               |> print
(1 x 1) TS with Dates.Date Index

 Index       x1
 Date        Float64
──────────────────────
 2007-01-01  0.768448

julia> ts[1, "x1"]

Base.join — Method

Joins/Column-binding

TS objects can be combined together column-wise using Index as the column key. There are four kinds of column-binding operations possible as of now. Each join operation works by performing a Set operation on the Index column and then merging the datasets based on the output from the Set operation. Each operation changes column names in the final object automatically if the operation encounters duplicate column names amongst the TS objects.

The following join types are supported:

join(ts1::TS, ts2::TS, ::JoinBoth)

a.k.a. inner join, takes the intersection of the indexes of ts1 and ts2, and then merges the columns of both the objects. The resulting object will only contain rows which are present in both the objects' indexes. The function will renamine the columns in the final object if they had same names in the TS objects.

join(ts1::TS, ts2::TS, ::JoinAll):

a.k.a. outer join, takes the union of the indexes of ts1 and ts2 before merging the other columns of input objects. The output will contain rows which are present in all the input objects while inserting missing values where a row was not present in any of the objects. This is the default behaviour if no JoinType object is provided.

join(ts1::TS, ts2::TS, ::JoinLeft):

Left join takes the index values which are present in the left object ts1 and finds matching index values in the right object ts2. The resulting object includes all the rows from the left object, the column values from the left object, and the values associated with matching index rows on the right. The operation inserts missing values where in the unmatched rows of the right object.

join(ts1::TS, ts2::TS, ::JoinRight)

Right join, similar to left join but works in the opposite direction. The final object contains all the rows from the right object while inserting missing values in rows missing from the left object.

The default behaviour is to assume JoinAll() if no JoinType object is provided to the join method.

cbind is an alias for join method.

Examples

julia> using Random;

julia> random(x) = rand(MersenneTwister(123), x);

julia> ts1 = TS(random(10), 1:10)

julia> ts2 = TS(random(10), 1:10)

julia> join(ts1, ts2, TSx.JoinAll())

julia> join(ts1, ts2);

julia> join(ts1, ts2, TSx.JoinBoth());

julia> join(ts1, ts2, TSx.JoinLeft());

julia> join(ts1, ts2, TSx.JoinRight());

julia> dates = collect(Date(2017,1,1):Day(1):Date(2017,1,10));

julia> ts1 = TS(random(length(dates)), dates) |> print
(10 x 1) TS with Dates.Date Index

 Index       x1
 Date        Float64
───────────────────────
 2017-01-01  0.768448
 2017-01-02  0.940515
 2017-01-03  0.673959
 2017-01-04  0.395453
 2017-01-05  0.313244
 2017-01-06  0.662555
 2017-01-07  0.586022
 2017-01-08  0.0521332
 2017-01-09  0.26864
 2017-01-10  0.108871

julia> dates = collect(Date(2017,1,1):Day(1):Date(2017,1,30));

julia> ts2 = TS(random(length(dates)), dates) |> print
(30 x 1) TS with Dates.Date Index

 Index       x1
 Date        Float64
───────────────────────
 2017-01-01  0.768448
 2017-01-02  0.940515
 2017-01-03  0.673959
 2017-01-04  0.395453
 2017-01-05  0.313244
 2017-01-06  0.662555
 2017-01-07  0.586022
 2017-01-08  0.0521332
     ⋮           ⋮
 2017-01-23  0.281066
 2017-01-24  0.792931
 2017-01-25  0.20923
 2017-01-26  0.918165
 2017-01-27  0.614255
 2017-01-28  0.802665
 2017-01-29  0.555668
 2017-01-30  0.940782
        14 rows omitted

julia> join(ts1, ts2) |> print

Base.names — Method

Column names

names(ts::TS)

Return a Vector{String} containing column names of the TS object, excludes index.

Examples

julia> names(TS([1:10 11:20]))
2-element Vector{String}:
 "x1"
 "x2"

Base.size — Method

Size methods

size(ts::TS) Return the number of rows and columns of ts as a tuple.

Examples

julia> TSx.size(TS([collect(1:100) collect(1:100) collect(1:100)]))
(100, 3)

Base.vcat — Method

vcat(ts::TS...; cols::Symbol=:setequal, source::Symbol=nothing)

vcat concatenates two or more arrays along dimension 1. This method implements the Base.vcat method.

The cols keyword argument determines the columns of the data frame :setequal: require all data frames to have the same column names disregarding order. If they appear in different orders, the order of the first provided data frame is used. :orderequal: require all data frames to have the same column names and in the same order. :intersect: only the columns present in all provided data frames are kept. If the intersection is empty, an empty data frame is returned. :union: columns present in at least one of the provided data frames are kept. Columns not present in some data frames are filled with missing where necessary.

The source keyword argument, if not nothing (the default), specifies the additional column to be added in the last position in the resulting data frame that will identify the source data frame.

Example

julia> using Random;

julia> random(x) = rand(MersenneTwister(123), x);

julia> dates1 = collect(Date(2017,1,1):Day(1):Date(2017,1,10));

julia> dates2 = collect(Date(2017,1,11):Day(1):Date(2017,1,30));

julia> ts1 = TS(randn(length(dates1)), dates1) |> print
(10 x 1) TS with Dates.Date Index

 Index       x1
 Date        Float64
────────────────────────
 2017-01-01  -0.420348
 2017-01-02   0.109363
 2017-01-03  -0.0702014
 2017-01-04   0.165618
 2017-01-05  -0.0556799
 2017-01-06  -0.147801
 2017-01-07  -2.50723
 2017-01-08  -0.099783
 2017-01-09   0.177526
 2017-01-10  -1.08461

julia> ts2 = TS(randn(length(dates2)), dates2) |> print
(20 x 1) TS with Dates.Date Index

 Index       x1
 Date        Float64
────────────────────────
 2017-01-11   2.15087
 2017-01-12   0.9203
 2017-01-13  -0.0879142
 2017-01-14  -0.930109
 2017-01-15   0.061117
 2017-01-16   0.0434627
 2017-01-17   0.0834733
 2017-01-18  -1.52281
     ⋮           ⋮
 2017-01-23  -0.756143
 2017-01-24   0.491623
 2017-01-25   0.549672
 2017-01-26   0.570689
 2017-01-27  -0.380011
 2017-01-28  -2.09965
 2017-01-29   1.37289
 2017-01-30  -0.462384
          4 rows omitted

julia> # vcat(ts1, ts2);        # FIXME

RecipesBase.apply_recipe — Function

Plottting

plot(ts::TS, cols::Vector{Int} = collect(1:TSx.ncol(ts))) plot(ts::TS, cols::Vector{T}) where {T<:Union{String, Symbol}} plot(ts::TS, cols::T) where {T<:Union{Int, String, Symbol}}

Plots a TS object with the index on the x-axis and selected cols on the y-axis. By default, plot all the columns. Columns can be selected using Int indexes, String(s), or Symbol(s).

Example

julia> dates = Date("2022-01-01"):Month(1):Date("2022-01-01")+Month(11);

julia> df = DataFrame(Index = dates, 
        val1 = randn(12), 
        val2 = randn(12),
        val3 = randn(12));

julia> TS(df)           |> print
(12 x 3) TS with Date Index

 Index       val1       val2       val3       
 Date        Float64    Float64    Float64    
──────────────────────────────────────────────
 2022-01-01  -2.2798    -0.801856  -0.696722
 2022-02-01  -0.266063   0.268206   0.743648
 2022-03-01   1.01839    1.64479   -0.296677
 2022-04-01   2.05753   -1.21628    0.339607
 2022-05-01   0.967545   1.21421    0.893949
 2022-06-01   0.191933  -0.56957   -1.0605
 2022-07-01  -1.73478    0.283089   0.360917
 2022-08-01  -0.305464  -1.55139    0.0505422
 2022-09-01  -2.40175    0.503355   0.743701
 2022-10-01  -0.273021   1.41333   -0.0682414
 2022-11-01   0.469246  -0.765097  -0.0119999
 2022-12-01  -1.12906    0.625611   1.72409

julia> using Plots

julia> plot(ts)

julia> plot(ts[1:6], [:val1, :val3])

julia> plot(ts, [1, 2], size=(600, 400))

TSx.apply — Method

Apply/Period conversion

apply(ts::TS, period::Union{T,Type{T}}, fun::V, index_at::Function=first) where {T<:Union{DatePeriod,TimePeriod}, V<:Function}

Apply fun to ts object based on period and return correctly indexed rows. This method is used for doing aggregation over a time period or to convert ts into an object of lower frequency (ex. from daily series to monthly).

period is any of Period types in the Dates module. Conversion from lower to a higher frequency will throw an error as interpolation isn't currently handled by this method.

By default, the method uses the first value of the index within the period to index the resulting aggregated object. This behaviour can be controlled by index_at argument which can take first or last as an input.

Examples

julia> using Random, Statistics;
julia> random(x) = rand(MersenneTwister(123), x);
julia> dates = collect(Date(2017,1,1):Day(1):Date(2018,3,10));

julia> df = DataFrame(Index = dates, x1 = random(length(dates)))
julia> ts = TS(df) |> print

julia> apply(ts, Month, first) |> print
(15 x 1) TS with Date Index

 Index       x1_first  
 Date        Float64   
───────────────────────
 2017-01-01  0.768448
 2017-02-01  0.790201
 2017-03-01  0.467219
 2017-04-01  0.783473
 2017-05-01  0.651354
 2017-06-01  0.373346
 2017-07-01  0.83296
 2017-08-01  0.132716
 2017-09-01  0.27899
 2017-10-01  0.995414
 2017-11-01  0.214132
 2017-12-01  0.832917
 2018-01-01  0.0409471
 2018-02-01  0.720163
 2018-03-01  0.87459

# alternate months
julia> apply(ts, Month(2), first) |> print
(8 x 1) TS with Date Index

 Index       x1_first  
 Date        Float64   
───────────────────────
 2017-01-01  0.768448
 2017-03-01  0.467219
 2017-05-01  0.651354
 2017-07-01  0.83296
 2017-09-01  0.27899
 2017-11-01  0.214132
 2018-01-01  0.0409471
 2018-03-01  0.87459

julia> ts_weekly = apply(ts, Week, Statistics.std) # weekly standard deviation
julia> show(ts_weekly)

julia> ts_weekly = apply(ts, Week, Statistics.std, last) # indexed by last date of the week
julia> show(ts_weekly)

TSx.index — Method

Index column

Return the index vector from the coredata DataFrame.

Examples

julia> using Random;

julia> random(x) = rand(MersenneTwister(123), x);

julia> ts = TS(random(10), Date("2022-02-01"):Month(1):Date("2022-02-01")+Month(9));

julia> show(ts)
(10 x 1) TS with Dates.Date Index

 Index       x1
 Date        Float64
───────────────────────
 2022-02-01  0.768448
 2022-03-01  0.940515
 2022-04-01  0.673959
 2022-05-01  0.395453
 2022-06-01  0.313244
 2022-07-01  0.662555
 2022-08-01  0.586022
 2022-09-01  0.0521332
 2022-10-01  0.26864
 2022-11-01  0.108871

julia> index(ts)
10-element Vector{Date}:
 2022-02-01
 2022-03-01
 2022-04-01
 2022-05-01
 2022-06-01
 2022-07-01
 2022-08-01
 2022-09-01
 2022-10-01
 2022-11-01

julia>  eltype(index(ts))
Date

TSx.lag — Function

Lagging

lag(ts::TS, lag_value::Int = 1)

Lag the ts object by the specified lag_value. The rows corresponding to lagged values will be rendered as missing. Negative values of lag are also accepted (see TSx.lead).

Examples

julia> using Random, Statistics;

julia> random(x) = rand(MersenneTwister(123), x);

julia> dates = collect(Date(2017,1,1):Day(1):Date(2017,1,10));

julia> ts = TS(DataFrame(Index = dates, x1 = random(length(dates)))) |> print
(10 x 1) TS with Dates.Date Index

 Index       x1
 Date        Float64
───────────────────────
 2017-01-01  0.768448
 2017-01-02  0.940515
 2017-01-03  0.673959
 2017-01-04  0.395453
 2017-01-05  0.313244
 2017-01-06  0.662555
 2017-01-07  0.586022
 2017-01-08  0.0521332
 2017-01-09  0.26864
 2017-01-10  0.108871

julia> l = lag(ts)
julia> show(l)

julia> l = lag(ts, 2) # Lags by 2 values;
julia> show(l)

TSx.lead — Function

Leading

lead(ts::TS, lead_value::Int = 1)

Similar to lag, this method leads the ts object by lead_value. The lead rows are inserted with missing. Negative values of lead are also accepted (see TSx.lag).

Examples

julia> using Random, Statistics;

julia> random(x) = rand(MersenneTwister(123), x);

julia> dates = collect(Date(2017,1,1):Day(1):Date(2018,3,10));

julia> ts = TS(DataFrame(Index = dates, x1 = random(length(dates)))) |> print
(434 x 1) TS with Dates.Date Index

 Index       x1
 Date        Float64
───────────────────────
 2017-01-01  0.768448
 2017-01-02  0.940515
 2017-01-03  0.673959
 2017-01-04  0.395453
 2017-01-05  0.313244
 2017-01-06  0.662555
 2017-01-07  0.586022
 2017-01-08  0.0521332
     ⋮           ⋮
 2018-03-03  0.127635
 2018-03-04  0.147813
 2018-03-05  0.873555
 2018-03-06  0.486486
 2018-03-07  0.495525
 2018-03-08  0.64075
 2018-03-09  0.375126
 2018-03-10  0.0338698
       418 rows omitted

julia> lead(ts) |> print # Leads once

julia> lead(ts, 2) # Leads by 2 values

TSx.log — Function

Log Function

log(ts::TS, complex::Bool = false)

This method computes the log value of the non-index columns in the TS object.

If the complex argument is true the function returns the log of negative numbers as complex numbers. But this also coerces the log of positive values as complex numbers with the imaginary component equal to 0.

TSx.ncol — Method

Size methods

ncol(ts::TS)

Return the number of columns of ts.

Examples

julia> using Random;

julia> random(x) = rand(MersenneTwister(123), x);

julia> TSx.ncol(TS([random(100) random(100) random(100)]))
3

TSx.nrow — Method

Size methods

nrow(ts::TS) Return the number of rows of ts.

Examples

julia> ts = TS(collect(1:10))

julia> TSx.nrow(ts)
10

TSx.rollapply — Method

Rolling Functions

rollapply(fun::Function, ts::TS, column::Any, windowsize::Int)

Apply a function to a column of ts for each continuous set of rows of size windowsize. column could be any of the DataFrame column selectors.

The output is a TS object with (nrow(ts) - windowsize + 1) rows indexed with the last index value of each window.

This method uses RollingFunctions package to implement this functionality.

Examples

julia> ts = TS(1:12, Date("2022-02-01"):Month(1):Date("2022-02-01")+Month(11))

julia> show(ts)
(12 x 1) TS with Dates.Date Index

 Index       x1
 Date        Int64
───────────────────
 2022-02-01      1
 2022-03-01      2
 2022-04-01      3
 2022-05-01      4
 2022-06-01      5
 2022-07-01      6
 2022-08-01      7
 2022-09-01      8
 2022-10-01      9
 2022-11-01     10
 2022-12-01     11
 2023-01-01     12

julia> rollapply(sum, ts, :x1, 10) |> print
(3 x 1) TS with Dates.Date Index

 Index       x1_rolling_sum
 Date        Float64
────────────────────────────
 2022-11-01            55.0
 2022-12-01            65.0
 2023-01-01            75.0

julia> rollapply(Statistics.mean, ts, 1, 5) |> print
(8 x 1) TS with Dates.Date Index

 Index       x1_rolling_mean
 Date        Float64
─────────────────────────────
 2022-06-01              3.0
 2022-07-01              4.0
 2022-08-01              5.0
 2022-09-01              6.0
 2022-10-01              7.0
 2022-11-01              8.0
 2022-12-01              9.0
 2023-01-01             10.0