TSx

Documentation for TSx.

TSx.TS

Base.getindex

Base.join

Base.size

TSx.apply

TSx.index

TSx.ncol

TSx.nrow

TSx.TS — Type

struct TS
  coredata :: DataFrame
end

::TS - A type to hold ordered data with an index.

A TS object is essentially a DataFrame with a specific column marked as an index and has the name Index. The DataFrame is sorted using index values during construction.

Constructors

TS(coredata::DataFrame, index::Union{String, Symbol, Int}=1)
TS(coredata::DataFrame, index::AbstractVector{T}) where {T<:Int}
TS(coredata::DataFrame, index::UnitRange{Int})
TS(coredata::AbstractVector{T}, index::AbstractVector{V}) where {T, V}
TS(coredata::AbstractVector{T}) where {T}
TS(coredata::AbstractArray{T,2}, meta::Dict=Dict{String, Any}()) where {T}

Examples

julia> df = DataFrame(x1 = randn(10))
julia> TS(df)

julia> df = DataFrame(Index = [1, 2, 3], x1 = randn(3))
julia> TS(df, 1)

julia> dates = collect(Date(2017,1,1):Day(1):Date(2017,1,10))
julia> df = DataFrame(dates = dates, x1 = randn(10))
julia> TS(df, :dates)
julia> TS(DataFrame(x1=randn(10), dates))

julia> TS(randn(10))
julia> TS(randn(10), dates)

Base.getindex — Method

Subsetting/Indexing

TS can be subset using row and column indices. The row selector could be an integer, a range, an array or it could also be a Date object or an ISO-formatted date string ("2007-04-10"). There are methods to subset on year, year-month, and year-quarter.

The latter two subset coredata by matching on the index column.

Column selector could be an integer or any other selector which DataFrame indexing supports. To fetch the index column one can use the index() method on the TS object.

Examples

julia> ts = TS([randn(10) randn(10) randn(10)])
julia> ts[1]
julia> ts[1:5]
julia> ts[1:5, 2]
julia> ts[1:5, 2:3]
julia> ts[[1, 9]]               # individual rows

julia> dates = collect(Date(2007):Day(1):Date(2008, 2, 22))
julia> ts = TS(randn(length(dates)), dates)
julia> ts[Date(2007, 01, 01)]
julia> ts[Date(2007)]
julia> ts[Year(2007)]
julia> ts[Year(2007), Month(11)]
julia> ts[Year(2007), Quarter(2)]
julia> ts["2007-01-01"]

julia> ts[1, :x1]
julia> ts[1, "x1"]

Base.join — Method

Joins/Column-binding

TS objects can be combined together column-wise using Index as the column key. There are four kinds of column-binding operations possible as of now. Each join operation works by performing a Set operation on the Index column and then merging the datasets based on the output from the Set operation. Each operation changes column names in the final object automatically if the operation encounters duplicate column names amongst the TS objects.

The following join types are supported:

join(ts1::TS, ts2::TS, ::JoinBoth)

a.k.a. inner join, takes the intersection of the indexes of ts1 and ts2, and then merges the columns of both the objects. The resulting object will only contain rows which are present in both the objects' indexes. The function will renamine the columns in the final object if they had same names in the TS objects.

join(ts1::TS, ts2::TS, ::JoinAll):

a.k.a. outer join, takes the union of the indexes of ts1 and ts2 before merging the other columns of input objects. The output will contain rows which are present in all the input objects while inserting missing values where a row was not present in any of the objects. This is the default behaviour if no JoinType object is provided.

join(ts1::TS, ts2::TS, ::JoinLeft):

Left join, takes the index values which are present in the left object ts1 and finds matching index values in the right object ts2. The resulting object includes all the rows from the left object, the column values from the left object, and the values associated with matching index rows on the right. The operation inserts missing values where in the unmatched rows of the right object.

join(ts1::TS, ts2::TS, ::JoinRight)

Right join, similar to left join but works in the opposite direction. The final object contains all the rows from the right object while inserting missing values in rows missing from the left object.

The default behaviour is to assume JoinAll() if no JoinType object is provided to the join method.

cbind is an alias for join method.

Examples

julia> ts1 = TS(randn(10), 1:10)
julia> ts2 = TS(randn(10), 1:10)

julia> join(ts1, ts2, JoinAll()) # with `missing` inserted
julia> join(ts1, ts2)            # same as JoinAll()
julia> join(ts1, ts2, JoinBoth())
julia> join(ts1, ts2, JoinLeft())
julia> join(ts1, ts2, JoinRight())

# Using TimeType objects
julia> dates = collect(Date(2017,1,1):Day(1):Date(2017,1,10))
julia> ts1 = TS(randn(length(dates)), dates)
julia> dates = collect(Date(2017,1,1):Day(1):Date(2017,1,30))
julia> ts2 = TS(randn(length(dates)), dates)

julia> join(ts1, ts2)

Base.size — Method

Size methods

size(ts::TS) Return the number of rows and columns of ts as a tuple.

Examples

julia> ts = TS([randn(100) randn(100) randn(100)])
julia> size(ts)

TSx.apply — Method

Apply/Period conversion

apply(ts::TS, period::Union{T,Type{T}}, fun::V, index_at::Function=first) where {T<:Union{DatePeriod,TimePeriod}, V<:Function}

Apply fun to ts object based on period and return correctly indexed rows. This method is used for doing aggregation over a time period or to convert ts into an object of lower frequency (ex. from daily series to monthly).

period is any of Period types in the Dates module. Conversion from lower to a higher frequency will throw an error as interpolation isn't currently handled by this method.

By default, the method uses the first value of the index within the period to index the resulting aggregated object. This behaviour can be controlled by index_at argument which can take first or last as an input.

Examples

julia> dates = collect(Date(2017,1,1):Day(1):Date(2018,3,10))
julia> ts = TS(DataFrame(Index = dates, x1 = randn(length(dates))))

# take the first observation in each month
julia> ts_monthly = apply(tsd, Month, first)
# alternate months
julia> ts_two_monthly = apply(tsd, Month(2), first)

# weekly standard deviation
julia> ts_monthly = apply(tsd, Week, Statistics.std)
# indexed by last date of the week
julia> ts_monthly = apply(tsd, Week, Statistics.std, last)

TSx.index — Method

Index column

Return the index vector from the TS DataFrame.

Examples

julia> ts = TS(randn(10), today():Month(1):today()+Month(9))
julia> index(ts)
julia> typeof(index(ts))

TSx.ncol — Method

Size methods

ncol(ts::TS) Return the number of columns of ts.

Examples

julia> ts = TS([randn(100) randn(100) randn(100)])
julia> ncol(ts)

TSx.nrow — Method

Size methods

nrow(ts::TS) Return the number of rows of ts.

Examples

julia> ts = TS(randn(100))
julia> nrow(ts)